move computeOptimalNWorkUnits to utils
